rafactor table view
