+++ console/src/state/tasks.rs
@@ -1,3 +1,4 @@
+use super::{Field, Metadata};
 use crate::{util::Percentage, view, warnings::Linter};
 use console_api as proto;
 use hdrhistogram::Histogram;
@@ -5,33 +6,24 @@ use std::{
     cell::RefCell,
     collections::HashMap,
     convert::{TryFrom, TryInto},
-    fmt,
-    io::Cursor,
     rc::{Rc, Weak},
     sync::Arc,
     time::{Duration, SystemTime},
 };
-use tui::{
-    style::{Color, Modifier},
-    text::Span,
-};
+use tui::{style::Color, text::Span};
 
 #[derive(Default, Debug)]
-pub(crate) struct State {
+pub(crate) struct TasksState {
     tasks: HashMap<u64, Rc<RefCell<Task>>>,
-    metas: HashMap<u64, Metadata>,
-    linters: Vec<Linter<Task>>,
-    last_updated_at: Option<SystemTime>,
     new_tasks: Vec<TaskRef>,
-    current_task_details: DetailsRef,
-    temporality: Temporality,
-    retain_for: Option<Duration>,
+    pub(crate) linters: Vec<Linter<Task>>,
 }
 
-#[derive(Debug)]
-enum Temporality {
-    Live,
-    Paused,
+#[derive(Debug, Default)]
+pub(crate) struct Details {
+    pub(crate) task_id: u64,
+    pub(crate) poll_times_histogram: Option<Histogram<u64>>,
+    pub(crate) last_updated_at: Option<SystemTime>,
 }
 
 #[derive(Debug, Copy, Clone)]
@@ -55,37 +47,21 @@ pub(crate) enum TaskState {
 }
 
 pub(crate) type TaskRef = Weak<RefCell<Task>>;
-pub(crate) type DetailsRef = Rc<RefCell<Option<Details>>>;
 
 #[derive(Debug)]
 pub(crate) struct Task {
     id: u64,
     fields: Vec<Field>,
     formatted_fields: Vec<Vec<Span<'static>>>,
-    stats: Stats,
+    stats: TaskStats,
     target: Arc<str>,
     name: Option<Arc<str>>,
     /// Currently active warnings for this task.
     warnings: Vec<Linter<Task>>,
 }
 
-#[derive(Debug, Default)]
-pub(crate) struct Details {
-    task_id: u64,
-    poll_times_histogram: Option<Histogram<u64>>,
-    last_updated_at: Option<SystemTime>,
-}
-
 #[derive(Debug)]
-pub(crate) struct Metadata {
-    field_names: Vec<Arc<str>>,
-    target: Arc<str>,
-    id: u64,
-    //TODO: add more metadata as needed
-}
-
-#[derive(Debug)]
-struct Stats {
+struct TaskStats {
     polls: u64,
     created_at: SystemTime,
     dropped_at: Option<SystemTime>,
@@ -110,36 +86,7 @@ struct Stats {
     self_wakes: u64,
 }
 
-#[derive(Debug)]
-pub(crate) struct Field {
-    pub(crate) name: Arc<str>,
-    pub(crate) value: FieldValue,
-}
-
-#[derive(Debug)]
-pub(crate) enum FieldValue {
-    Bool(bool),
-    Str(String),
-    U64(u64),
-    I64(i64),
-    Debug(String),
-}
-
-impl State {
-    pub(crate) fn with_retain_for(mut self, retain_for: Option<Duration>) -> Self {
-        self.retain_for = retain_for;
-        self
-    }
-
-    pub(crate) fn with_linters(mut self, linters: impl IntoIterator<Item = Linter<Task>>) -> Self {
-        self.linters.extend(linters.into_iter());
-        self
-    }
-
-    pub(crate) fn last_updated_at(&self) -> Option<SystemTime> {
-        self.last_updated_at
-    }
-
+impl TasksState {
     /// Returns any new tasks that were added since the last task update.
     pub(crate) fn take_new_tasks(&mut self) -> impl Iterator<Item = TaskRef> + '_ {
         self.new_tasks.drain(..)
@@ -149,28 +96,16 @@ impl State {
         &mut self,
         styles: &view::Styles,
         update: proto::tasks::TaskUpdate,
-        new_metadata: Option<proto::RegisterMetadata>,
-        now: Option<SystemTime>,
+        metas: &HashMap<u64, Metadata>,
+        clear_new_list: bool,
     ) {
-        if let Some(now) = now {
-            self.last_updated_at = Some(now);
-        }
-
-        if let Some(new_metadata) = new_metadata {
-            let metas = new_metadata.metadata.into_iter().filter_map(|meta| {
-                let id = meta.id?.id;
-                let metadata = meta.metadata?;
-                Some((id, Metadata::from_proto(metadata, id)))
-            });
-            self.metas.extend(metas);
-        }
-
         let mut stats_update = update.stats_update;
         let new_list = &mut self.new_tasks;
-        new_list.clear();
+        if clear_new_list {
+            new_list.clear();
+        }
         let linters = &self.linters;
 
-        let metas = &mut self.metas;
         let new_tasks = update.new_tasks.into_iter().filter_map(|mut task| {
             if task.id.is_none() {
                 tracing::warn!(?task, "skipping task with no id");
@@ -233,67 +168,32 @@ impl State {
         }
     }
 
-    pub(crate) fn details_ref(&self) -> DetailsRef {
-        self.current_task_details.clone()
-    }
+    pub(crate) fn retain_active(&mut self, now: SystemTime, retain_for: Duration) {
+        self.tasks.retain(|_, task| {
+            let task = task.borrow();
 
-    pub(crate) fn update_task_details(&mut self, update: proto::tasks::TaskDetails) {
-        if let Some(id) = update.task_id {
-            let details = Details {
-                task_id: id.id,
-                poll_times_histogram: update.poll_times_histogram.and_then(|data| {
-                    hdrhistogram::serialization::Deserializer::new()
-                        .deserialize(&mut Cursor::new(&data))
-                        .ok()
-                }),
-                last_updated_at: update.now.map(|now| now.try_into().unwrap()),
-            };
-
-            *self.current_task_details.borrow_mut() = Some(details);
-        }
-    }
-
-    pub(crate) fn unset_task_details(&mut self) {
-        *self.current_task_details.borrow_mut() = None;
-    }
-
-    pub(crate) fn retain_active(&mut self) {
-        // Don't clean up stopped tasks while the console is paused.
-        if self.is_paused() {
-            return;
-        }
-
-        if let (Some(now), Some(retain_for)) = (self.last_updated_at, self.retain_for) {
-            self.tasks.retain(|_, task| {
-                let task = task.borrow();
-
-                task.stats
-                    .dropped_at
-                    .map(|d| {
-                        let dropped_for = now.duration_since(d).unwrap();
-                        retain_for > dropped_for
-                    })
-                    .unwrap_or(true)
-            })
-        }
-    }
-
-    // temporality methods
-
-    pub(crate) fn pause(&mut self) {
-        self.temporality = Temporality::Paused;
-    }
-
-    pub(crate) fn resume(&mut self) {
-        self.temporality = Temporality::Live;
-    }
-
-    pub(crate) fn is_paused(&self) -> bool {
-        matches!(self.temporality, Temporality::Paused)
+            task.stats
+                .dropped_at
+                .map(|d| {
+                    let dropped_for = now.duration_since(d).unwrap();
+                    retain_for > dropped_for
+                })
+                .unwrap_or(true)
+        })
     }
 
     pub(crate) fn warnings(&self) -> impl Iterator<Item = &Linter<Task>> {
         self.linters.iter().filter(|linter| linter.count() > 0)
+    }
+}
+
+impl Details {
+    pub(crate) fn task_id(&self) -> u64 {
+        self.task_id
+    }
+
+    pub(crate) fn poll_times_histogram(&self) -> Option<&Histogram<u64>> {
+        self.poll_times_histogram.as_ref()
     }
 }
 
@@ -430,17 +330,7 @@ impl Task {
     }
 }
 
-impl Details {
-    pub(crate) fn task_id(&self) -> u64 {
-        self.task_id
-    }
-
-    pub(crate) fn poll_times_histogram(&self) -> Option<&Histogram<u64>> {
-        self.poll_times_histogram.as_ref()
-    }
-}
-
-impl From<proto::tasks::Stats> for Stats {
+impl From<proto::tasks::Stats> for TaskStats {
     fn from(pb: proto::tasks::Stats) -> Self {
         fn pb_duration(dur: prost_types::Duration) -> Duration {
             let secs =
@@ -480,34 +370,6 @@ impl From<proto::tasks::Stats> for Stats {
     }
 }
 
-impl Metadata {
-    fn from_proto(pb: proto::Metadata, id: u64) -> Self {
-        Self {
-            field_names: pb.field_names.into_iter().map(|n| n.into()).collect(),
-            target: pb.target.into(),
-            id,
-        }
-    }
-}
-
-impl From<proto::field::Value> for FieldValue {
-    fn from(pb: proto::field::Value) -> Self {
-        match pb {
-            proto::field::Value::BoolVal(v) => Self::Bool(v),
-            proto::field::Value::StrVal(v) => Self::Str(v),
-            proto::field::Value::I64Val(v) => Self::I64(v),
-            proto::field::Value::U64Val(v) => Self::U64(v),
-            proto::field::Value::DebugVal(v) => Self::Debug(v),
-        }
-    }
-}
-
-impl Default for Temporality {
-    fn default() -> Self {
-        Self::Live
-    }
-}
-
 impl Default for SortBy {
     fn default() -> Self {
         Self::Total
@@ -542,6 +404,12 @@ impl SortBy {
     }
 }
 
+impl view::SortBy for SortBy {
+    fn as_column(&self) -> usize {
+        *self as usize
+    }
+}
+
 impl TryFrom<usize> for SortBy {
     type Error = ();
     fn try_from(idx: usize) -> Result<Self, Self::Error> {
@@ -555,179 +423,6 @@ impl TryFrom<usize> for SortBy {
             idx if idx == Self::Idle as usize => Ok(Self::Idle),
             idx if idx == Self::Polls as usize => Ok(Self::Polls),
             _ => Err(()),
-        }
-    }
-}
-
-// === impl Field ===
-
-impl Field {
-    const SPAWN_LOCATION: &'static str = "spawn.location";
-    const NAME: &'static str = "task.name";
-
-    /// Converts a wire-format `Field` into an internal `Field` representation,
-    /// using the provided `Metadata` for the task span that the field came
-    /// from.
-    ///
-    /// If the field is invalid or it has a string value which is empty, this
-    /// returns `None`.
-    fn from_proto(
-        proto::Field {
-            name,
-            metadata_id,
-            value,
-        }: proto::Field,
-        meta: &Metadata,
-    ) -> Option<Self> {
-        use proto::field::Name;
-        let name: Arc<str> = match name? {
-            Name::StrName(n) => n.into(),
-            Name::NameIdx(idx) => {
-                let meta_id = metadata_id.map(|m| m.id);
-                if meta_id != Some(meta.id) {
-                    tracing::warn!(
-                        task.meta_id = meta.id,
-                        field.meta.id = ?meta_id,
-                        field.name_index = idx,
-                        ?meta,
-                        "skipping malformed field name (metadata id mismatch)"
-                    );
-                    debug_assert_eq!(
-                        meta_id,
-                        Some(meta.id),
-                        "malformed field name: metadata ID mismatch! (name idx={}; metadata={:#?})",
-                        idx,
-                        meta,
-                    );
-                    return None;
-                }
-                match meta.field_names.get(idx as usize).cloned() {
-                    Some(name) => name,
-                    None => {
-                        tracing::warn!(
-                            task.meta_id = meta.id,
-                            field.meta.id = ?meta_id,
-                            field.name_index = idx,
-                            ?meta,
-                            "missing field name for index"
-                        );
-                        return None;
-                    }
-                }
-            }
-        };
-
-        debug_assert!(
-            value.is_some(),
-            "missing field value for field `{:?}` (metadata={:#?})",
-            name,
-            meta,
-        );
-        let mut value = FieldValue::from(value?)
-            // if the value is an empty string, just skip it.
-            .ensure_nonempty()?;
-
-        if &*name == Field::SPAWN_LOCATION {
-            value = value.truncate_registry_path();
-        }
-
-        Some(Self { name, value })
-    }
-
-    fn make_formatted(styles: &view::Styles, fields: &mut Vec<Field>) -> Vec<Vec<Span<'static>>> {
-        use std::cmp::Ordering;
-
-        let key_style = styles.fg(Color::LightBlue).add_modifier(Modifier::BOLD);
-        let delim_style = styles.fg(Color::LightBlue).add_modifier(Modifier::DIM);
-        let val_style = styles.fg(Color::Yellow);
-
-        fields.sort_unstable_by(|left, right| {
-            if &*left.name == Field::NAME {
-                return Ordering::Less;
-            }
-
-            if &*right.name == Field::NAME {
-                return Ordering::Greater;
-            }
-
-            if &*left.name == Field::SPAWN_LOCATION {
-                return Ordering::Greater;
-            }
-
-            if &*right.name == Field::SPAWN_LOCATION {
-                return Ordering::Less;
-            }
-
-            left.name.cmp(&right.name)
-        });
-
-        let mut formatted = Vec::with_capacity(fields.len());
-        let mut fields = fields.iter();
-        if let Some(field) = fields.next() {
-            formatted.push(vec![
-                Span::styled(field.name.to_string(), key_style),
-                Span::styled("=", delim_style),
-                Span::styled(format!("{} ", field.value), val_style),
-            ]);
-            for field in fields {
-                formatted.push(vec![
-                    // Span::styled(", ", delim_style),
-                    Span::styled(field.name.to_string(), key_style),
-                    Span::styled("=", delim_style),
-                    Span::styled(format!("{} ", field.value), val_style),
-                ])
-            }
-        }
-        formatted
-    }
-}
-
-// === impl FieldValue ===
-
-impl fmt::Display for FieldValue {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            FieldValue::Bool(v) => fmt::Display::fmt(v, f)?,
-            FieldValue::Str(v) => fmt::Display::fmt(v, f)?,
-            FieldValue::U64(v) => fmt::Display::fmt(v, f)?,
-            FieldValue::Debug(v) => fmt::Display::fmt(v, f)?,
-            FieldValue::I64(v) => fmt::Display::fmt(v, f)?,
-        }
-
-        Ok(())
-    }
-}
-
-impl FieldValue {
-    /// Truncates paths including `.cargo/registry`.
-    fn truncate_registry_path(self) -> Self {
-        use once_cell::sync::OnceCell;
-        use regex::Regex;
-        use std::borrow::Cow;
-
-        static REGEX: OnceCell<Regex> = OnceCell::new();
-        let regex = REGEX.get_or_init(|| {
-            Regex::new(r#".*/\.cargo/registry/src/[^/]*/"#).expect("failed to compile regex")
-        });
-
-        let s = match self {
-            FieldValue::Str(s) | FieldValue::Debug(s) => s,
-            f => return f,
-        };
-
-        let s = match regex.replace(&s, "<cargo>/") {
-            Cow::Owned(s) => s,
-            // String was not modified, return the original.
-            Cow::Borrowed(_) => s,
-        };
-        FieldValue::Debug(s)
-    }
-
-    /// If `self` is an empty string, returns `None`. Otherwise, returns `Some(self)`.
-    fn ensure_nonempty(self) -> Option<Self> {
-        match self {
-            FieldValue::Debug(s) | FieldValue::Str(s) if s.is_empty() => None,
-            val => Some(val),
         }
     }
 }
